fix: listener and renderer type errors
