+++ libs/core/package.json
@@ -37,6 +37,7 @@
     "react": "~16.8.6"
   },
   "dependencies": {
+    "@types/stylenames": "^1.1.0",
     "css-to-style": "^1.2.1",
     "classnames": "^2.2.6",
     "stylenames": "^1.1.6",
