+++ libs/core/src/lib/components/wrapper-component.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-/// <reference path="../@types/geteventlisteners.d.ts" />
+/// <reference path="../renderer/geteventlisteners.ts" />
 
 import { AfterViewInit, ChangeDetectorRef, ElementRef, Input, NgZone, OnChanges, Renderer2, SimpleChanges, AfterContentInit, Directive } from '@angular/core';
 import classnames from 'classnames';
