+++ libs/core/tsconfig.lib.json
@@ -11,7 +11,6 @@
     "experimentalDecorators": true,
     "importHelpers": true,
     "types": [],
-    "typeRoots": ["./src/lib/@types"],
     "lib": ["dom", "es2017"]
   },
   "angularCompilerOptions": {
